BAU: Moved get-question AWS tests into their own directory and renameds file
